Refine progress state module boundaries
