+++ src/app/state/progress.rs
@@ -1,61 +1,12 @@
-/// Modal progress indicator for slow tasks.
-/// Identifies the long-running task responsible for updating the progress overlay.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum ProgressTaskKind {
-    /// Moving files to trash.
-    TrashMove,
-    /// Loading a waveform from disk.
-    WavLoad,
-    /// Scanning a sample source.
-    Scan,
-    /// Running background analysis jobs.
-    Analysis,
-    /// Normalizing audio samples.
-    Normalization,
-    /// Exporting waveform slice batches in the background.
-    SelectionExport,
-    /// Copying, moving, or restoring files in the background.
-    FileOps,
-    /// Filtering or rebuilding browser results in the background.
-    Search,
-}
+mod snapshots;
+mod task;
+
+pub use snapshots::{AnalysisProgressSnapshot, RunningJobSnapshot};
+pub use task::ProgressTaskKind;
 
 use std::collections::HashMap;
 use std::time::Instant;
-
-#[derive(Clone, Debug)]
-struct ProgressTaskState {
-    modal: bool,
-    title: String,
-    detail: Option<String>,
-    completed: usize,
-    total: usize,
-    cancelable: bool,
-    cancel_requested: bool,
-    last_update_at: Option<Instant>,
-    last_progress_at: Option<Instant>,
-    analysis: Option<AnalysisProgressSnapshot>,
-    started_at: Instant,
-}
-
-impl ProgressTaskState {
-    fn new(modal: bool, title: impl Into<String>, total: usize, cancelable: bool) -> Self {
-        let now = Instant::now();
-        Self {
-            modal,
-            title: title.into(),
-            detail: None,
-            completed: 0,
-            total,
-            cancelable,
-            cancel_requested: false,
-            last_update_at: Some(now),
-            last_progress_at: Some(now),
-            analysis: None,
-            started_at: now,
-        }
-    }
-}
+use task::{ProgressTaskState, task_priority};
 
 /// UI state for the progress overlay and its counters.
 #[derive(Clone, Debug, Default)]
@@ -86,58 +37,6 @@ pub struct ProgressOverlayState {
     pub analysis: Option<AnalysisProgressSnapshot>,
     /// Active footer-progress contenders across background task families.
     task_states: HashMap<ProgressTaskKind, ProgressTaskState>,
-}
-
-/// Snapshot of analysis progress for display.
-#[derive(Clone, Debug)]
-pub struct AnalysisProgressSnapshot {
-    /// Number of pending jobs.
-    pub pending: usize,
-    /// Number of running jobs.
-    pub running: usize,
-    /// Number of failed jobs.
-    pub failed: usize,
-    /// Completed samples count.
-    pub samples_completed: usize,
-    /// Total samples to process.
-    pub samples_total: usize,
-    /// Snapshot of running jobs.
-    pub running_jobs: Vec<RunningJobSnapshot>,
-    /// Staleness threshold in seconds.
-    pub stale_after_secs: Option<i64>,
-}
-
-/// Summary of a running job heartbeat for display.
-#[derive(Clone, Debug)]
-pub struct RunningJobSnapshot {
-    /// Human-readable job label.
-    pub label: String,
-    /// Last heartbeat timestamp, epoch seconds.
-    pub last_heartbeat_at: Option<i64>,
-    /// Whether the job appears stalled.
-    pub possibly_stalled: bool,
-}
-
-impl RunningJobSnapshot {
-    /// Build a snapshot and mark it stalled when the heartbeat is stale.
-    pub fn from_heartbeat(
-        label: String,
-        last_heartbeat_at: Option<i64>,
-        stale_after_secs: Option<i64>,
-        now_epoch: Option<i64>,
-    ) -> Self {
-        let possibly_stalled = match (last_heartbeat_at, stale_after_secs, now_epoch) {
-            (Some(heartbeat), Some(stale_after), Some(now)) => {
-                now.saturating_sub(heartbeat) >= stale_after
-            }
-            _ => false,
-        };
-        Self {
-            label,
-            last_heartbeat_at,
-            possibly_stalled,
-        }
-    }
 }
 
 impl ProgressOverlayState {
@@ -313,24 +212,32 @@ impl ProgressOverlayState {
         let previous = self.task;
         let selected = self.select_visible_task(previous);
         let Some(task) = selected else {
-            self.visible = false;
-            self.modal = false;
-            self.task = None;
-            self.title.clear();
-            self.detail = None;
-            self.completed = 0;
-            self.total = 0;
-            self.cancelable = false;
-            self.cancel_requested = false;
-            self.last_update_at = None;
-            self.last_progress_at = None;
-            self.analysis = None;
+            self.apply_hidden_projection();
             return;
         };
-        let slot = self
-            .task_states
-            .get(&task)
-            .expect("selected progress task should exist");
+        if let Some(slot) = self.task_states.get(&task).cloned() {
+            self.apply_visible_projection(task, &slot);
+        } else {
+            self.apply_hidden_projection();
+        }
+    }
+
+    fn apply_hidden_projection(&mut self) {
+        self.visible = false;
+        self.modal = false;
+        self.task = None;
+        self.title.clear();
+        self.detail = None;
+        self.completed = 0;
+        self.total = 0;
+        self.cancelable = false;
+        self.cancel_requested = false;
+        self.last_update_at = None;
+        self.last_progress_at = None;
+        self.analysis = None;
+    }
+
+    fn apply_visible_projection(&mut self, task: ProgressTaskKind, slot: &ProgressTaskState) {
         self.visible = true;
         self.modal = slot.modal;
         self.task = Some(task);
@@ -370,83 +277,5 @@ impl ProgressOverlayState {
     }
 }
 
-fn task_priority(task: ProgressTaskKind) -> u8 {
-    match task {
-        ProgressTaskKind::TrashMove => 100,
-        ProgressTaskKind::Scan => 90,
-        ProgressTaskKind::Analysis => 80,
-        ProgressTaskKind::FileOps => 70,
-        ProgressTaskKind::Normalization => 60,
-        ProgressTaskKind::SelectionExport => 50,
-        ProgressTaskKind::WavLoad => 20,
-        ProgressTaskKind::Search => 10,
-    }
-}
-
 #[cfg(test)]
-mod tests {
-    use super::{ProgressOverlayState, ProgressTaskKind, RunningJobSnapshot};
-
-    #[test]
-    fn progress_fraction_handles_zero_total() {
-        let progress = ProgressOverlayState::new(ProgressTaskKind::TrashMove, "Task", 0, false);
-        assert_eq!(progress.fraction(), 0.0);
-    }
-
-    #[test]
-    fn progress_reset_clears_visibility() {
-        let mut progress = ProgressOverlayState::new(ProgressTaskKind::TrashMove, "Task", 2, true);
-        progress.completed = 3;
-        assert!(progress.fraction() <= 1.0);
-        progress.reset();
-        assert!(!progress.visible);
-        assert_eq!(progress.task, None);
-        assert_eq!(progress.completed, 0);
-        assert_eq!(progress.total, 0);
-    }
-
-    #[test]
-    fn running_job_marks_stale_heartbeat() {
-        let snapshot =
-            RunningJobSnapshot::from_heartbeat("job".to_string(), Some(10), Some(5), Some(20));
-        assert!(snapshot.possibly_stalled);
-
-        let snapshot =
-            RunningJobSnapshot::from_heartbeat("job".to_string(), Some(18), Some(5), Some(20));
-        assert!(!snapshot.possibly_stalled);
-    }
-
-    #[test]
-    fn higher_priority_background_task_wins_footer_lane() {
-        let mut progress = ProgressOverlayState::default();
-        progress.show_task(ProgressTaskKind::Analysis, false, "Analyzing", 10, true);
-        progress.show_task(
-            ProgressTaskKind::WavLoad,
-            false,
-            "Loading samples",
-            0,
-            false,
-        );
-
-        assert_eq!(progress.task, Some(ProgressTaskKind::Analysis));
-        assert_eq!(progress.title, "Analyzing");
-    }
-
-    #[test]
-    fn clearing_visible_task_reveals_next_contender() {
-        let mut progress = ProgressOverlayState::default();
-        progress.show_task(ProgressTaskKind::Analysis, false, "Analyzing", 10, true);
-        progress.show_task(
-            ProgressTaskKind::WavLoad,
-            false,
-            "Loading samples",
-            0,
-            false,
-        );
-
-        progress.clear_task(ProgressTaskKind::Analysis);
-
-        assert_eq!(progress.task, Some(ProgressTaskKind::WavLoad));
-        assert_eq!(progress.title, "Loading samples");
-    }
-}
+mod tests;
