+++ src/app_core/native_bridge/tests/mod.rs
@@ -95,9 +95,9 @@ fn runtime_exit_returns_structured_shutdown_timing_once() {
 #[test]
 fn runtime_exit_detaches_active_controller_shutdown_work() {
     /// Long enough for the test file operation to prove shutdown does not wait on it.
-    const BLOCKING_FILE_OP: Duration = Duration::from_secs(15);
+    const BLOCKING_FILE_OP: Duration = Duration::from_secs(45);
     /// Upper bound for the detached runtime-exit handoff.
-    const DETACHED_EXIT_BUDGET: Duration = Duration::from_secs(10);
+    const DETACHED_EXIT_BUDGET: Duration = Duration::from_secs(20);
 
     let base = tempdir().expect("create temp config dir");
     let _base_guard = crate::app_dirs::ConfigBaseGuard::set(base.path().to_path_buf());
